// apps/api/src/prisma/contract.d.ts
export type CodecTypes = PgTypes;
// ...
export namespace Models {
  export type public_User = {
    id: CodecTypes['pg/text@1']['output'];
    name: CodecTypes['pg/text@1']['output'];
    email: CodecTypes['pg/text@1']['output'];
    timezone: CodecTypes['pg/text@1']['output'];
    createdAt: TimestampString<3>;
    projects: public_Project[];
    readonly [RelationKeys]?: 'projects';
  };
  export type public_Project = {
    id: CodecTypes['pg/text@1']['output'];
    userId: CodecTypes['pg/text@1']['output'];
    name: CodecTypes['pg/text@1']['output'];
    createdAt: TimestampString<3>;
    subjects: public_Subject[];
    user: public_User;
    readonly [RelationKeys]?: 'subjects' | 'user';
  };
  export type public_Subject = {
    id: CodecTypes['pg/text@1']['output'];
    projectId: CodecTypes['pg/text@1']['output'];
    name: CodecTypes['pg/text@1']['output'];
    kpiTarget: CodecTypes['pg/int4@1']['output'];
    kpiType: 'totalTime' | 'totalRepeat';
    periodType: 'day' | 'week' | 'twoWeek' | 'month';
    startDate: CodecTypes['pg/date-string@1']['output'];
    link: CodecTypes['pg/text@1']['output'] | null;
    createdAt: TimestampString<3>;
    archivedAt: TimestampString<3> | null;
    events: public_SubjectEvent[];
    kpiChanges: public_KpiChange[];
    logs: public_ProgressLog[];
    project: public_Project;
    readonly [RelationKeys]?: 'events' | 'kpiChanges' | 'logs' | 'project';
  };
  export type public_KpiChange = {
    id: CodecTypes['pg/text@1']['output'];
    subjectId: CodecTypes['pg/text@1']['output'];
    oldValue: CodecTypes['pg/int4@1']['output'];
    newValue: CodecTypes['pg/int4@1']['output'];
    createdAt: TimestampString<3>;
    subject: public_Subject;
    readonly [RelationKeys]?: 'subject';
  };
  export type public_ProgressLog = {
    id: CodecTypes['pg/text@1']['output'];
    subjectId: CodecTypes['pg/text@1']['output'];
    amount: CodecTypes['pg/int4@1']['output'];
    loggedAt: TimestampString<3>;
    note: CodecTypes['pg/text@1']['output'] | null;
    createdAt: TimestampString<3>;
    subject: public_Subject;
    readonly [RelationKeys]?: 'subject';
  };
  export type public_SubjectEvent = {
    id: CodecTypes['pg/text@1']['output'];
    subjectId: CodecTypes['pg/text@1']['output'];
    status: 'finish' | 'miss';
    periodStart: CodecTypes['pg/date-string@1']['output'];
    periodEnd: CodecTypes['pg/date-string@1']['output'];
    achieved: CodecTypes['pg/int4@1']['output'];
    kpiSnapshot: CodecTypes['pg/int4@1']['output'];
    createdAt: TimestampString<3>;
    subject: public_Subject;
    readonly [RelationKeys]?: 'subject';
  };
}